Added some unfinished and unused classes
Item is intended to be a unifying base class for items, and should allow
for transitioning a block from an inventory item, being placed in a
grid, being removed from the grid into the world, and picked up from the
world into the players inventory.

Belt is the hotbar for items.